made a proper ev class
